pyproject.toml
[tool.poetry]
name = "django-ninja-fk-bug"
version = "0.1.0"
description = "Reproduce bug with fk have to be int"
authors = ["M.R. Sopacua <144725145+msopacua@users.noreply.github.com>"]
license = "Public Domain"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-ninja = "<1"
cuid2 = "^2.0.0"
dateparser = "^1.2.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.17.2"
django-stubs = "^4.2.6"
pytest = "^7.4.3"
pytest-django = "^4.7.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.settings"
python_files = ["tests.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
